Add support for a command wrapper for the app
This should mean that it's easy enough to pipx install the project and then
there's a nice handy command to run. I've chosen `mdex` (MarkDown EXplorer),
but of course that's totally up to you.

This PR is more to suggest making this just a wee bit easier to run so it
becomes a habit. :-)